Add ChainStateProof related methods
